exectrace: Add --time-offset-file option
This reads the mtime of the given file and shifts all timestamps
from perf.data so that the time of the last sample aligns with it.

This is a hack that lets us correlate with realtime clock events
even though we have recorded events with another clock, such as
CLOCK_MONOTONIC.

Using the realtime clock directly when capturing is not supported for
the events we need. And even if it was, using the realtime clock would
risk introducing skew between samples.

(Relying on mtime is not nice. But perf's own "header" display reads
"captured on" from the file stat. If it's good enough for perf, it
ought to be good enough for us, for the time being.)